requirements.txt
# Core
torch>=2.4.0
transformers>=4.49.0
accelerate>=0.33.0
safetensors>=0.4.3
sentencepiece>=0.2.0
pillow>=10.3.0

# Qwen2.5-VL utilities (Qwen team's official helper)
qwen-vl-utils>=0.0.10

# Analysis / detector / figures
scikit-learn>=1.5.0
numpy>=1.26.0
scipy>=1.13.0
pandas>=2.2.0
matplotlib>=3.9.0
seaborn>=0.13.0

# Data loaders + tqdm
requests>=2.32.0
tqdm>=4.66.0
datasets>=2.19.0

# Optional: only needed if running the stretch-goal HITs cross-reference
# (output_attentions=True on Qwen2.5-VL)
einops>=0.8.0
